requirements.txt
--index-url https://download.pytorch.org/whl/cpu
flask==2.3.2
torch==2.3.0+cpu
diffusers==0.29.2
transformers==4.31.0
accelerate==0.21.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
gunicorn==21.2.0